+++ src/index.js
@@ -1,6 +1,7 @@
 import Resolver from '@forge/resolver';
 import ForgeUI, { render } from '@forge/ui';
 import api, {route} from '@forge/api'
+import {getMotivationRatings} from './motivation';
 
 const resolver = new Resolver();
 
@@ -12,10 +13,6 @@ resolver.define('getText', async (req) => {
 
     return 'Hello, world!!!!!!';
 });
-
-
-import EditMotivation from './Motivation/Edit';
-import ViewMotivation from './Motivation/View';
 
 resolver.define('getMotivation', async (req) => {
     const motivationsCount = await getMotivationRatings(req);
@@ -46,30 +43,5 @@ const getPerformanceRatings = async function(req) {
     
     return issuePerformances;
 }
-
-const getMotivationRatings = async function(req) {
-    var motivationsCount = {};
-    for (var issue of data.issues) {
-        var issueMotivationField = issue.fields.customfield_10048;
-        for (const property in issueMotivationField) {
-            if (property != 'myMotivationRating' && issueMotivationField[`${property}`]) {
-                const previous = motivationsCount[`${property}`] ? motivationsCount[`${property}`] : 0;
-                motivationsCount[`${property}`] = previous + 1;
-            }
-        }
-    }
-
-    return motivationsCount;
-}
-
-export const renderViewMotivation = render(
-    <ViewMotivation/>
-);
-
-export const renderEditMotivation = render(
-    <EditMotivation/>
-);
-
-
 
 export const handler = resolver.getDefinitions();
