+++ static/manager-dashboard/src/App.js
@@ -1,31 +1,9 @@
-import React, { useEffect, useState } from 'react';
-import { invoke } from '@forge/bridge';
-import RadarChart from './shared/RadarChart';
+import Motivation from './Components/Motivation';
 
 function App() {
-    const [data, setData] = useState(null);
-
-    useEffect(() => {
-        invoke('getMotivation', { example: 'my-invoke-variable' }).then(setData);
-    }, []);
-
-    const labels = [];
-    const dataPoints = [];
-    for (const property in data) {
-        labels.push(property);
-        dataPoints.push(data[`${property}`]);
-    }
-    const borderColour = 'rgba(0, 82, 204)'
-    const backgroundColour = 'rgba(179, 212, 255)'
     return (
         <div>
-            <RadarChart
-                name='Motivation'
-                labels={labels}
-                dataPoints={dataPoints}
-                borderColour={borderColour}
-                backgroundColour={backgroundColour}
-            />
+            <Motivation/>
         </div>
     );
 }
